refactor(layout): extract line construction
Move text-span line clustering and segment splitting into a dedicated layout module so the layout entrypoint coordinates the pipeline at a higher level.
